Refactor listgroup to list_group for consitency with input_group
