+++ docs/components_page/components/list_group/links.py
@@ -2,12 +2,12 @@ import dash_bootstrap_components as dbc
 import dash_html_components as html
 from dash.dependencies import Input, Output
 
-listgroup = html.Div(
+list_group = html.Div(
     [
         dbc.ListGroup(
             [
                 dbc.ListGroupItem(
-                    "Internal link", href="/l/components/listgroup"
+                    "Internal link", href="/l/components/list_group"
                 ),
                 dbc.ListGroupItem("External link", href="https://google.com"),
                 dbc.ListGroupItem(
