+++ docs/components_page/components/list_group/simple.py
@@ -1,6 +1,6 @@
 import dash_bootstrap_components as dbc
 
-listgroup = dbc.ListGroup(
+list_group = dbc.ListGroup(
     [
         dbc.ListGroupItem("Item 1"),
         dbc.ListGroupItem("Item 2"),
